server/requirements.txt
fastapi
uvicorn
bcrypt
pyjwt
upstash-redis